+++ app/components/UI/Bridge/hooks/useTokenSelection.ts
@@ -6,10 +6,12 @@ import {
   setDestToken,
   selectSourceToken,
   selectDestToken,
+  selectDestAmount,
   setIsDestTokenManuallySet,
 } from '../../../../core/redux/slices/bridge';
 import { BridgeToken, TokenSelectorType } from '../types';
 import { useSwitchTokens } from './useSwitchTokens';
+import { useIsNetworkEnabled } from './useIsNetworkEnabled';
 
 /**
  * Hook to manage token selection logic for Bridge token selector
@@ -21,7 +23,9 @@ export const useTokenSelection = (type: TokenSelectorType) => {
   const navigation = useNavigation();
   const sourceToken = useSelector(selectSourceToken);
   const destToken = useSelector(selectDestToken);
+  const destAmount = useSelector(selectDestAmount);
   const { handleSwitchTokens } = useSwitchTokens();
+  const isDestNetworkEnabled = useIsNetworkEnabled(destToken?.chainId);
 
   const handleTokenPress = useCallback(
     async (token: BridgeToken) => {
@@ -35,8 +39,21 @@ export const useTokenSelection = (type: TokenSelectorType) => {
         token.chainId === otherToken.chainId;
 
       if (isSelectingOtherToken && sourceToken && destToken) {
+        // Only allow swap if the destination network (which would become source) is enabled
+        if (!isDestNetworkEnabled) {
+          // Cannot swap - dest network is disabled, just go back
+          navigation.goBack();
+          return;
+        }
+
         // Swap the tokens: old source becomes dest, old dest becomes source
-        await handleSwitchTokens()();
+        // Pass destAmount so it becomes the new sourceAmount after swap
+        try {
+          await handleSwitchTokens(destAmount)();
+        } catch {
+          // Network switch failed - still navigate back but state may be inconsistent
+          // The user can retry from the main view
+        }
       } else {
         // Normal selection: just update the current token
         dispatch(isSourcePicker ? setSourceToken(token) : setDestToken(token));
@@ -47,7 +64,16 @@ export const useTokenSelection = (type: TokenSelectorType) => {
 
       navigation.goBack();
     },
-    [type, sourceToken, destToken, dispatch, navigation, handleSwitchTokens],
+    [
+      type,
+      sourceToken,
+      destToken,
+      destAmount,
+      dispatch,
+      navigation,
+      handleSwitchTokens,
+      isDestNetworkEnabled,
+    ],
   );
 
   const selectedToken =
